requirements.txt
# Core Django
Django==5.0.6
djangorestframework==3.15.1

# Static files serving
whitenoise==6.6.0

# Django extensions
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.7
dj-database-url==2.2.0

# Production server
gunicorn==22.0.0

# AI Integration
openai==1.35.3
anthropic==0.25.9
httpx==0.25.2

# Utilities
python-dotenv==1.0.1
pytz==2024.1
Pillow==10.3.0

# Authentication
django-otp==1.4.0
qrcode==7.4.2
argon2-cffi==23.1.0
bcrypt==4.1.2

# Task queue
celery==5.3.6
redis==5.0.4
django-celery-beat==2.8.1

# AWS Integration
boto3==1.34.131
django-storages==1.14.3

# Development tools (optional)
django-debug-toolbar==4.4.6
